backend/requirements.txt
# Core NLP and ML libraries
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Text processing and analysis
textblob==0.17.1
gensim==4.3.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Web framework for API
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1 